requirements.txt
discord.py==2.2.3
ndjson==0.3.1
python-dotenv==1.0.0
